Fix cases where entities are re-added with different members.
A recent patch added the SocialService class. This class existed
previously, but with different members. As a result, the removed members
of the previous version were incorrectly displayed as if they were
present in the current version.

This commit adds logic to handle the case where an entity is re-added.
If the entity associated with an added element has a previous version of
the element, then each member of the previous version is compared to
those of the new version. If a member is not present in the new
version, then it is retroactively removed by setting its removed state,
and adding, to the member entity, the patch that removes the parent
entity. This turns out to work well with the updates page, which lists
removed members along with a removed parent.

Since a member entity may now include patches that apply to the parent
entity, such patches will be incorrectly repeated when rendering the
parent's patches as a history list. This is corrected by adding an
optional filter to MergePatches. Patches on the right-side are merged
only if the condition of the given filter is met.