Improve code blocks to work with themes.
This removes the syntax highlighting CSS generated by Chroma. Instead,
the same CSS is written into doc.css with any coloring replaced by
variables. This is less flexible with Chroma, which may decide to change
its formatting in the future. But it's already not flexible enough to
support themes, so there's a trade-off. I'd rather have themes.