In security field patches, icon displays previous state.
The information is derived from a "patch" structure, which contains the
current state of the item plus the information needed to transform it
into the next state.

Attempting to display the next state would get weird for properties,
where the display is derived from two separate fields that get patched
separately. To do it correctly would require knowledge of the "final"
next state, which is too much for such a trivial feature.