Fix rbxdhist jobs beings skipped.
The format of jobs was recently changed to include a git hash. A part of
this change was that the status that accompanies a job no longer
occurs.

This is resolved by skipping the check for status if the job has a git
hash. A possibly more stable alternative worth considering is to use
another epoch.

Bumps github.com/robloxapi/rbxdhist to v0.3.0.